+++ experiments/tokio-graceful-shutdown/src/main.rs
@@ -1,11 +1,10 @@
 use std::time::Duration;
 
+use lib_wc::sync::Shutdown;
 use tokio::select;
 use tokio::signal::ctrl_c;
 use tokio::sync::{broadcast, mpsc};
 use tokio::time::{interval, interval_at};
-
-mod shutdown;
 
 /// Graceful Shutdown (https://tokio.rs/tokio/topics/shutdown)
 ///
@@ -27,7 +26,7 @@ async fn main() {
 
     // Spawn 10 tens
     for i in 0..10 {
-        let shutdown = shutdown::Shutdown::new(notify_shutdown.subscribe());
+        let shutdown = Shutdown::new(notify_shutdown.subscribe());
         let shutdown_complete_tx = shutdown_complete_tx.clone();
         tokio::spawn(async move {
             let mut task = Task::new(i, shutdown, shutdown_complete_tx);
@@ -70,18 +69,14 @@ struct Task {
     id: u32,
 
     /// Shutdown signal that is used to signal that the task should stop
-    shutdown: shutdown::Shutdown,
+    shutdown: Shutdown,
 
     /// Implicitly used to signal that the task has finished by being dropped
     _shutdown_complete_tx: mpsc::Sender<()>,
 }
 
 impl Task {
-    pub fn new(
-        id: u32,
-        shutdown: shutdown::Shutdown,
-        _shutdown_complete_tx: mpsc::Sender<()>,
-    ) -> Self {
+    pub fn new(id: u32, shutdown: Shutdown, _shutdown_complete_tx: mpsc::Sender<()>) -> Self {
         Self {
             id,
             shutdown,
