+++ src/concurrent/sync/mod.rs
@@ -2,7 +2,9 @@
 
 pub use rate_limiter::RateLimiter;
 pub mod ds;
+pub use shutdown::Shutdown;
 mod rate_limiter;
+mod shutdown;
 
 cfg_dangerous! {
     pub use channels::{mpmc, oneshot};
