+++ src/lib.rs
@@ -5,6 +5,7 @@
 //! # Concurrency Tools
 //!
 //! * [`sync::RateLimiter`], a client-side rate limiter
+//! * [`sync::Shutdown`], a tool to signal shutdown and wait for completion of asynchronous tasks
 //! * [`sync::ds::BasicSharedMap`], a concurrent map that can be cloned and shared between threads
 //! * [`executors::RayonThreadPool`], a thread pool which can wait for all tasks to complete before shutting down
 //!
